+++ src/media/image/sixel.rs
@@ -4,14 +4,8 @@
 //!
 //! [Sixel]: https://en.wikipedia.org/wiki/Sixel
 //
-// NOTE: doc features don't show from /all/
 
-#[cfg(all(feature = "alloc", feature = "term"))]
-#[cfg(any(feature = "io", feature = "std"))]
-#[cfg(any(feature = "dep_hashbrown", feature = "std"))]
-pub mod legacy;
-
-crate::structural_mods! { // _mods, _pub_mods, _hidden
+crate::structural_mods! { // _mods, _hidden
     _mods {
         // re-exports
         #[cfg_attr(nightly_doc, doc(cfg(feature = "image")))]
@@ -25,17 +19,6 @@ crate::structural_mods! { // _mods, _pub_mods, _hidden
             #[cfg_attr(nightly_doc, doc(cfg(feature = "__dbg")))]
             pub use static: sixel_encoder:64+64 in devela_base_core
         } }
-    }
-    _pub_mods {
-        #[doc(inline)]
-        #[cfg(all(feature = "alloc", feature = "term"))]
-        #[cfg(any(feature = "io", feature = "std"))]
-        #[cfg(any(feature = "dep_hashbrown", feature = "std"))]
-        // FIX
-        #[cfg_attr(nightly_doc, doc(cfg(all(feature = "alloc", feature = "term"))))]
-        #[cfg_attr(nightly_doc, doc(cfg(any(feature = "io", feature = "std"))))]
-        #[cfg_attr(nightly_doc, doc(cfg(any(feature = "dep_hashbrown", feature = "std"))))]
-        pub use super::legacy::*;
     }
     _hidden {
         crate::__dbg! {